+++ src/test/java/next/controller/LoginControllerTest.java
@@ -2,7 +2,7 @@ package next.controller;
 
 import core.mvc.view.ForwardView;
 import core.mvc.view.ModelAndView;
-import core.mvc.tobe.DispatcherServlet;
+import core.mvc.asis.DispatcherServlet;
 import core.mvc.tobe.AnnotationHandlerMapping;
 import core.mvc.tobe.HandlerExecution;
 import org.junit.jupiter.api.BeforeEach;
